extract instant export to separate class
